+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [v1.55.0] - 2025-06-17
 
 - Promote `buf beta stats` to `buf stats`.
 - Update built-in Well-Known Types to Protobuf v31.1.
@@ -1379,7 +1379,7 @@ buf check breaking proto --against .git#branch=master,subdir=proto
 
 Initial beta release.
 
-[Unreleased]: https://github.com/bufbuild/buf/compare/v1.54.0...HEAD
+[v1.55.0]: https://github.com/bufbuild/buf/compare/v1.54.0...v1.55.0
 [v1.54.0]: https://github.com/bufbuild/buf/compare/v1.53.0...v1.54.0
 [v1.53.0]: https://github.com/bufbuild/buf/compare/v1.52.1...v1.53.0
 [v1.52.1]: https://github.com/bufbuild/buf/compare/v1.52.0...v1.52.1